Added a --use-tidy and a --force option.
The --use-tidy option is to prettify the output. I can't get ERB
to do this, I tried Haml but that just introduced other problems.
I decided to leave it up to `tidy` as it works well in the specs.

The --force option is to stop files being destroyed/overwritten,
whether maliciously or by accident.